Mass rename, namespace refactoring
