+++ FalconBMS Alternative Launcher Cs/App.xaml
@@ -1,7 +1,7 @@
-﻿<Application x:Class="FalconBMS_Alternative_Launcher_Cs.App"
+﻿<Application x:Class="FalconBMS.Launcher.App"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             StartupUri="MainWindow.xaml">
+             StartupUri="Windows/MainWindow.xaml">
     <Application.Resources>
         <ResourceDictionary>
             <ResourceDictionary.MergedDictionaries>
