+++ FalconBMS Alternative Launcher Cs/AppProperties.cs
@@ -1,6 +1,8 @@
 ﻿using System.Windows;
 
-namespace FalconBMS_Alternative_Launcher_Cs
+using FalconBMS.Launcher.Windows;
+
+namespace FalconBMS.Launcher
 {
     public class AppProperties
     {
