+++ FalconBMS Alternative Launcher Cs/Input/KeyAssgn.cs
@@ -1,12 +1,10 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
+
+using FalconBMS.Launcher.Windows;
 
 using Microsoft.DirectX.DirectInput;
 
-namespace FalconBMS_Alternative_Launcher_Cs
+namespace FalconBMS.Launcher.Input
 {
     public class KeyAssgn
     {
