+++ FalconBMS Alternative Launcher Cs/Launcher.cs
@@ -1,7 +1,9 @@
 ﻿using System.IO;
 using System.Windows;
 
-namespace FalconBMS_Alternative_Launcher_Cs
+using FalconBMS.Launcher.Windows;
+
+namespace FalconBMS.Launcher
 {
     public class Launcher
     {
