+++ FalconBMS Alternative Launcher Cs/Windows/KeyMappingWindow.xaml
@@ -1,9 +1,8 @@
-﻿<Controls:MetroWindow x:Class="FalconBMS_Alternative_Launcher_Cs.KeyMappingWindow"
+﻿<Controls:MetroWindow x:Class="FalconBMS.Launcher.Windows.KeyMappingWindow"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-        xmlns:local="clr-namespace:FalconBMS_Alternative_Launcher_Cs"
         xmlns:Controls="clr-namespace:MahApps.Metro.Controls;assembly=MahApps.Metro"
         mc:Ignorable="d"
         GlowBrush="Black" BorderThickness="0"
